feat: Add `posts` app to INSTALLED_APPS
The code changes include adding the `'posts'` app to the `INSTALLED_APPS` list in the `settings.py` file. This allows the Django project to recognize and use the `posts` app.

This commit adds the `'posts'` app to the `INSTALLED_APPS` list, enabling the project to utilize the functionality provided by the `posts` app.