feat: Add 'likes_count' field to PostSerializer and order posts by likes count
This commit adds a new field, 'likes_count', to the PostSerializer in the 'posts' app. The 'likes_count' field is an IntegerField that represents the total number of likes for a post. Additionally, the 'get_posts' view in the 'posts' app is modified to order the posts by the likes count in descending order. This change improves the user experience by displaying the most popular posts first.