feat: Add 'comments' and 'comments_count' fields to PostSerializer
This commit modifies the PostSerializer in the 'posts' app to include the 'comments' and 'comments_count' fields. The 'comments' field is a nested serializer that represents the comments associated with a post, and the 'comments_count' field is an IntegerField that represents the total number of comments for a post. This change enhances the functionality of the serializer and allows for the display of comments along with the post information.