feat: Update URLs in 'add.bru' and 'remove.bru' files
This commit updates the URLs in the 'add.bru' and 'remove.bru' files in the 'django/comments' and 'django/likes' directories. The URLs are changed to use the correct endpoints for adding and removing comments and likes. This change ensures that the correct API endpoints are used for these actions, improving the functionality and reliability of the application.