feat: Update 'ger user roles.bru' to use GET request for retrieving user roles
This commit updates the 'ger user roles.bru' file in the 'django' directory to use a GET request instead of a POST request for retrieving user roles. The URL has been changed to '{{base_url}}/users/roles/get' and the authentication requirement has been set to 'none'. This change enhances the functionality of the application by aligning the request method with the intended purpose of retrieving data.