feat: Add 'get_all' endpoint for areas and roles
This commit adds the 'get_all' endpoint for areas and roles in the backend. The endpoints are implemented in the 'area/views.py' and 'roles/views.py' files, allowing authenticated users to retrieve all areas and roles respectively. This addition enhances the functionality of the application by providing users with access to all available areas and roles.